Update user association to events and navigation
Corrected the association between events and users by moving the event
id out of the users table and creating a creator_id in the events table.

The events index also now has a link to a users events page (although I
still need to update the users show page to show the events).